Move multiaccessor to fix circular reference
